Fix chat imports via esm.sh
