+++ README.md
@@ -41,4 +41,4 @@ For a guided setup experience, visit `/wizard`. This multi-step form walks throu
 
 ### Telegram-style Scheduler
 
-Navigate to `/chat` for a lightweight chat interface. Enter a group ID and type messages from multiple AI users or yourself. Each message can be assigned a specific time using `datetime-local` inputs. When you click **Schedule All**, the page sends the collected messages to `/advanced` where they are stored and later dispatched by the cron job. Click **Reply** next to any message to mark the next one as a response.
+Navigate to `/chat` for a Telegram-style interface powered by **react-chat-elements**. Enter a group ID and sender name, type your message and optionally pick a time. Messages appear in a scrolling chat window and can be scheduled in bulk with **Schedule All** which posts them to `/advanced`.
