+++ package.json
@@ -10,6 +10,9 @@
   "author": "",
   "license": "ISC",
   "type": "commonjs",
+  "dependencies": {
+    "react-chat-elements": "^12.0.18"
+  },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20250605.0",
     "hono": "^4.7.11",
